fix(types): Resolve all mypy strict type checking errors
Fixes 20 mypy type errors across 4 files to achieve 100% type safety:

Schema enhancements:
- Add DocumentQualityScore and OCRRoutingRecommendation type aliases
- Provide backward compatibility for routing module imports

Type annotations:
- Fix numpy type inference in layout_lite.py by explicit float casting
- Cast blur_score, noise_std, and gray_std to resolve union type issues

Missing functions:
- Add calculate_dqs() to aggregate page-level IQA metrics
- Add calculate_pre_ocr_risk() for OCR difficulty prediction
- Implement proper DQS calculation and risk scoring logic

Field corrections:
- Fix PageLayoutSummary field names (page_number, complexity_score)
- Remove non-existent PageAttributes nesting
- Add missing teacher_iqa and teacher_usage fields across modules

All 35 source files now pass mypy strict type checking with zero errors.

+++ src/image_preprocessing_detector/detection/layout_lite.py
@@ -499,7 +499,7 @@ def detect_fuzzy_scan(
 
     # Normalize blur score (inverse: lower variance = more blur)
     # Typical range: sharp images have variance >500, blurry <100
-    blur_score = 1.0 - min(1.0, laplacian_var / 500.0)
+    blur_score = float(1.0 - min(1.0, laplacian_var / 500.0))
 
     # Estimate noise using high-frequency components
     # Apply high-pass filter (difference from Gaussian blur)
@@ -507,7 +507,7 @@ def detect_fuzzy_scan(
     noise_image = cv2.absdiff(gray, blurred)
 
     # Calculate noise metric as standard deviation of noise image
-    noise_std = np.std(noise_image)
+    noise_std = float(np.std(noise_image))
 
     # Normalize noise score
     # Typical range: clean images have std <10, noisy >30
@@ -590,7 +590,7 @@ def detect_watermark(
 
     # Calculate opacity score from intensity variations
     # Watermarks typically have semi-transparent, uniform intensity
-    gray_std = np.std(gray)
+    gray_std = float(np.std(gray))
     opacity_score = 1.0 - min(1.0, gray_std / 50.0)  # Normalize
 
     # Detection logic: high low-frequency energy indicates watermark
